+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/spawners/from_files/from_files.py
@@ -14,7 +14,7 @@ import omni.kit.commands
 from pxr import Gf, Sdf, Usd
 
 from omni.isaac.orbit.sim import converters, schemas
-from omni.isaac.orbit.sim.utils import bind_physics_material, bind_visual_material, clone
+from omni.isaac.orbit.sim.utils import bind_physics_material, bind_visual_material, clone, select_usd_variants
 
 if TYPE_CHECKING:
     from . import from_files_cfg
@@ -232,6 +232,10 @@ def _spawn_from_usd_file(
         )
     else:
         carb.log_warn(f"A prim already exists at prim path: '{prim_path}'.")
+
+    # modify variants
+    if hasattr(cfg, "variants") and cfg.variants is not None:
+        select_usd_variants(prim_path, cfg.variants)
 
     # modify rigid body properties
     if cfg.rigid_props is not None: