+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/utils.py
@@ -777,3 +777,82 @@ def find_global_fixed_joint_prim(
                 break
 
     return fixed_joint_prim
+
+
+"""
+USD Variants.
+"""
+
+
+def select_usd_variants(prim_path: str, variants: object | dict[str, str], stage: Usd.Stage | None = None):
+    """Sets the variant selections from the specified variant sets on a USD prim.
+
+    `USD Variants`_ are a very powerful tool in USD composition that allows prims to have different options on
+    a single asset. This can be done by modifying variations of the same prim parameters per variant option in a set.
+    This function acts as a script-based utility to set the variant selections for the specified variant sets on a
+    USD prim.
+
+    The function takes a dictionary or a config class mapping variant set names to variant selections. For instance,
+    if we have a prim at ``"/World/Table"`` with two variant sets: "color" and "size", we can set the variant
+    selections as follows:
+
+    .. code-block:: python
+
+        select_usd_variants(
+            prim_path="/World/Table",
+            variants={
+                "color": "red",
+                "size": "large",
+            },
+        )
+
+    Alternatively, we can use a config class to define the variant selections:
+
+    .. code-block:: python
+
+        @configclass
+        class TableVariants:
+            color: Literal["blue", "red"] = "red"
+            size: Literal["small", "large"] = "large"
+
+        select_usd_variants(
+            prim_path="/World/Table",
+            variants=TableVariants(),
+        )
+
+    Args:
+        prim_path: The path of the USD prim.
+        variants: A dictionary or config class mapping variant set names to variant selections.
+        stage: The USD stage. Defaults to None, in which case, the current stage is used.
+
+    Raises:
+        ValueError: If the prim at the specified path is not valid.
+
+    .. _USD Variants: https://graphics.pixar.com/usd/docs/USD-Glossary.html#USDGlossary-Variant
+    """
+    # Resolve stage
+    if stage is None:
+        stage = stage_utils.get_current_stage()
+    # Obtain prim
+    prim = stage.GetPrimAtPath(prim_path)
+    if not prim.IsValid():
+        raise ValueError(f"Prim at path '{prim_path}' is not valid.")
+    # Convert to dict if we have a configclass object.
+    if not isinstance(variants, dict):
+        variants = variants.to_dict()
+
+    existing_variant_sets = prim.GetVariantSets()
+    for variant_set_name, variant_selection in variants.items():
+        # Check if the variant set exists on the prim.
+        if not existing_variant_sets.HasVariantSet(variant_set_name):
+            carb.log_warn(f"Variant set '{variant_set_name}' does not exist on prim '{prim_path}'.")
+            continue
+
+        variant_set = existing_variant_sets.GetVariantSet(variant_set_name)
+        # Only set the variant selection if it is different from the current selection.
+        if variant_set.GetVariantSelection() != variant_selection:
+            variant_set.SetVariantSelection(variant_selection)
+            carb.log_info(
+                f"Setting variant selection '{variant_selection}' for variant set '{variant_set_name}' on"
+                f" prim '{prim_path}'."
+            )